[ANDROSDK-1968] move fields and service to network layer
